+++ src/test/java/com/wrapper/spotify/requests/authorization/authorization_code/AuthorizationCodeRequestTest.java
@@ -30,7 +30,7 @@ public class AuthorizationCodeRequestTest {
       final AuthorizationCodeCredentials tokens = request.post();
       assertEquals("BQBY2M94xNVE_7p7x1MhNd2I1UNs62cv-CVDXkDwh5YqSiKJceKRXwJfUrLmJFKO7GfiCZKTh8oEEj3b84bZx1Qy52qwGYCVhX6yHPJY4VDday-hC1YMPOWyIt9Bp05UuJb673btr6T1YOd0DliheWDyqQ", tokens.getAccessToken());
       assertEquals("Bearer", tokens.getTokenType());
-      assertEquals(3600, tokens.getExpiresIn());
+      assertEquals(3600, (int) tokens.getExpiresIn());
       assertEquals("AQAZ54v-sV7LO_R64q76KtDMKeQcPkBIPAuKFqYr1kSAeaU8_S8ZxbnqcNizeQiSJr5DhMsJvCdgS7_KUrHd7rw1z7h_FJkL5OVOnthZrNFdO5NL7gUvNJRF6hdbIkAnEHM", tokens.getRefreshToken());
     } catch (Exception e) {
       fail(e.getMessage());
