+++ src/test/java/com/wrapper/spotify/requests/data/browse/GetListOfFeaturedPlaylistsRequestTest.java
@@ -47,9 +47,9 @@ public class GetListOfFeaturedPlaylistsRequestTest {
 
         Paging<PlaylistSimplified> playlistPage = featuredPlaylists.getPlaylists();
 
-        assertEquals(23, playlistPage.getTotal());
-        assertEquals(0, playlistPage.getOffset());
-        assertEquals(1, playlistPage.getLimit());
+        assertEquals(23, (int) playlistPage.getTotal());
+        assertEquals(0, (int) playlistPage.getOffset());
+        assertEquals(1, (int) playlistPage.getLimit());
         assertEquals("https://api.spotify.com/v1/browse/featured-playlists?timestamp=2017-12-03T01%3A40%3A49&offset=1&limit=1",
                 playlistPage.getNext());
         assertNull(playlistPage.getPrevious());
@@ -94,9 +94,9 @@ public class GetListOfFeaturedPlaylistsRequestTest {
 
     Paging<PlaylistSimplified> playlistPage = featuredPlaylists.getPlaylists();
 
-    assertEquals(23, playlistPage.getTotal());
-    assertEquals(0, playlistPage.getOffset());
-    assertEquals(1, playlistPage.getLimit());
+    assertEquals(23, (int) playlistPage.getTotal());
+    assertEquals(0, (int) playlistPage.getOffset());
+    assertEquals(1, (int) playlistPage.getLimit());
     assertEquals("https://api.spotify.com/v1/browse/featured-playlists?timestamp=2017-12-03T01%3A40%3A49&offset=1&limit=1",
             playlistPage.getNext());
     assertNull(playlistPage.getPrevious());
